requirements.txt
chardet==3.0.4
psutil==5.1.3
requests==2.18.1
steamfiles==0.1.3
watchdog==0.8.3
python-valve==0.2.1
